move reslayer to utils
